cleaned up the basic strucutre, added delegating constructors to Blade. in the middle of implementing StimulationSteward

+++ signals/ConstantSignal.cpp
@@ -9,7 +9,7 @@
 #include "ConstantSignal.hpp"
 
 
-ConstantSignal::ConstantSignal(int magnitude, int startTime, int endTime, int length) : Signal(false){
+ConstantSignal::ConstantSignal(float magnitude, int startTime, int endTime, int length) : Signal(false){
     this->magnitude = magnitude;
     generateFullSignal(startTime, endTime, length);
 }
